refactor #19: import 경로변경 및 변수 리네이밍밍
